Refactor details package name to edit
