+++ app/src/main/java/com/lambdaschool/hackathon_portal/di/FragmentComponent.kt
@@ -4,7 +4,7 @@ import androidx.fragment.app.Fragment
 import com.lambdaschool.hackathon_portal.ui.fragments.account.AccountFragment
 import com.lambdaschool.hackathon_portal.ui.fragments.create.CreateHackathonFragment
 import com.lambdaschool.hackathon_portal.ui.fragments.dashboard.DashboardFragment
-import com.lambdaschool.hackathon_portal.ui.fragments.details.EditHackathonFragment
+import com.lambdaschool.hackathon_portal.ui.fragments.edit.EditHackathonFragment
 import com.lambdaschool.hackathon_portal.ui.fragments.login.LoginFragment
 import com.lambdaschool.hackathon_portal.ui.fragments.login.LoginModule
 import com.lambdaschool.hackathon_portal.ui.fragments.settings.SettingsFragment
