+++ app/src/main/res/layout/fragment_edit_hackathon.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.fragments.details.EditHackathonFragment">
+    tools:context=".ui.fragments.edit.EditHackathonFragment">
 
     <EditText
         android:id="@+id/fragment_edit_hackathon_edit_text_hackathon_name"
